refactor: split shell dialogs and debug flip-province helpers (#3878)
Extract leader-selection setup-field mixins (header/options/footer),
new-game setup progress vs error dialog parts, production breakdown
layout constants, and flip-province target-resolution helpers.
