+++ app/lib/features/shell/new_game_leader_selection_dialog_setup_fields_options.dart
@@ -1,77 +1,9 @@
 part of 'new_game_leader_selection_dialog.dart';
 
-/// Seed, advanced-start, infinite-mode, terrain, and footer fields for
-/// [NewGameLeaderSelectionDialog] (Refs #3878 shell modularization).
-mixin _NewGameLeaderSelectionDialogSetupFields
+/// Advanced-start, infinite-mode, and terrain fields for
+/// [NewGameLeaderSelectionDialog] (Refs #3878).
+mixin _NewGameLeaderSelectionDialogSetupFieldsOptions
     on State<NewGameLeaderSelectionDialog>, _NewGameLeaderSelectionDialogStateBase {
-  Widget _buildHeader(AppLocalizations l10n, _LeaderDialogTextStyles styles) {
-    // Mockup header order (DLG10001 `.dialog-body`): centered title, centered
-    // italic intro, then the brass divider beneath both.
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Text(
-          l10n.shell_leaderDialog_title,
-          key: const ValueKey<String>('leaderSelectionDialogTitle'),
-          style: styles.title,
-          textAlign: TextAlign.center,
-        ),
-        const SizedBox(height: CtSpacing.xs),
-        Text(
-          l10n.shell_leaderDialog_intro,
-          key: const ValueKey<String>('leaderSelectionDialogIntro'),
-          style: styles.intro,
-          textAlign: TextAlign.center,
-        ),
-        const SizedBox(height: CtSpacing.ml),
-        const CtBrassDivider(
-          key: ValueKey<String>('leaderSelectionDialogBrassDivider'),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildSeedField(
-    ThemeData theme,
-    AppLocalizations l10n,
-    _LeaderDialogTextStyles styles,
-  ) {
-    final OutlineInputBorder idleBorder = OutlineInputBorder(
-      borderRadius: BorderRadius.zero,
-      borderSide: BorderSide(color: EditorialMonoclePalette.border),
-    );
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Text(l10n.shell_leaderDialog_seedLabel, style: styles.fieldLabel),
-        const SizedBox(height: CtSpacing.m / 2),
-        TextField(
-          controller: _seedController,
-          keyboardType: TextInputType.number,
-          style: theme.textTheme.bodyMedium?.copyWith(
-            color: EditorialMonoclePalette.fg,
-          ),
-          decoration: InputDecoration(
-            isDense: true,
-            border: idleBorder,
-            enabledBorder: idleBorder,
-            focusedBorder: OutlineInputBorder(
-              borderRadius: BorderRadius.zero,
-              borderSide: BorderSide(
-                color: EditorialMonoclePalette.accent,
-                width: 2,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(height: CtSpacing.s),
-        Text(l10n.shell_leaderDialog_seedHelper, style: styles.helper),
-      ],
-    );
-  }
-
   bool get _advancedStartEnabled => widget.baseConfig.isLockedFullInitProfile;
 
   String _advancedStartLabel(AppLocalizations l10n, AdvancedStartType type) {
@@ -172,52 +104,6 @@ mixin _NewGameLeaderSelectionDialogSetupFields
         ),
       ],
     );
-  }
-
-  Widget _buildFooterButtons(AppLocalizations l10n, BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      children: [
-        CtNinePatchButton(
-          onPressed: widget.onCancel,
-          child: Text(l10n.common_cancel),
-        ),
-        const SizedBox(width: CtSpacing.m),
-        CtNinePatchButton(
-          onPressed: _startEnabled ? () => _handleStartPressed(context) : null,
-          enabled: _startEnabled,
-          child: Text(l10n.common_start),
-        ),
-      ],
-    );
-  }
-
-  void _handleStartPressed(BuildContext context) {
-    final seed = NewGameLeaderSelectionDialog.parseSeedInput(
-      _seedController.text,
-    );
-    Navigator.of(context).pop();
-    widget.onConfirmed(
-      List<String>.from(_orderedGpIdsBySlot),
-      Map<String, String>.from(_leaderByGpId),
-      seed,
-      _infiniteMode,
-      _terrainVariation,
-      _aiProfileByGpIdForCallback(),
-      _advancedStartEnabled ? _advancedStart : AdvancedStartType.none,
-    );
-  }
-
-  Map<String, String?> _aiProfileByGpIdForCallback() {
-    final out = <String, String?>{};
-    for (var slot = 1; slot < _kNumSlots; slot++) {
-      final gpId = _orderedGpIdsBySlot[slot];
-      final profileName = _profileBySlot[slot];
-      if (profileName != null && profileName.isNotEmpty) {
-        out[gpId] = profileName;
-      }
-    }
-    return out;
   }
 
   Widget _buildTerrainVariationField(
